requirements.txt
# Silero VAD
--index-url https://download.pytorch.org/whl/cu129
torch==2.8.0+cu129
--index-url https://pypi.org/simple/
numpy==2.3.5
pydub==0.25.1
packaging==25.0

# Visualize
librosa==0.11.0
matplotlib==3.10.7

# Open AI Whispher
openai-whisper==20250625

#
Resemblyzer==0.1.4